941_Problem
For every element check if next element is greater than current. For each such pair , increment count. Check if count is zero of length of array. In both case return False. Now check for every next pair, if next element is smaller than current and increment count. At the end,  compare if count is same as length of array.