234_Problem
We can work for palindrome with string with two pointers. So we need to convert the linked list to string. Then we will run the logic of palindrome. So we take two pointer, one at start and one at end. Keep comparing data of these pointers and increment and decrement respective pointers.