21_Problem
Since we have two linked list to combine, obviously we need to keep traversing. So we need two pointer, one at the start of merged list and at the end of merged list. While both input lists are alive, just check smallest of them and add it to merged list at end. At the end, when one of the list is exhausted, just add remaining second list to the end of merged list