708_Problem
430	05/11/2022	Linked List, Recursion	"You are given a doubly linked list, which contains nodes that have a next pointer, a previous pointer, and an additional child pointer. This child pointer may or may not point to a separate doubly linked list, also containing these special nodes. These child lists may have one or more children of their own, and so on, to produce a multilevel data structure as shown in the example below.
Given the head of the first level of the list, flatten the list so that all the nodes appear in a single-level, doubly linked list. Let curr be a node with a child list. The nodes in the child list should appear after curr and before curr.next in the flattened list.
Return the head of the flattened list. The nodes in the list must have all of their child pointers set to null."	Start from head. For every node , check if node has a child list. If yes, simply disconnect it as child add it after current node. Make sure that after adding child list as next, all 4 pointers viz curr's next, child list's prev, child list's next and curr next's prev are properly set. As We would be traversing with child node the recursive way weill take care of all children list inside of child list.