138_Problem
We want to create complete new list with all new nodes. So first create every node and assigned it to next of original node from which it is created. Now for every random pointer in original list, assign the next pointer of random to next of current node. Now for node, next of that node points to next of next of next node. Return the new list's head node.